Extract ElementBase base class from Element
